+++ linux/src/config.h
@@ -1,0 +1,7 @@
+#ifndef BRLASER_CONFIG_H
+#define BRLASER_CONFIG_H
+
+#define PACKAGE "brlaser"
+#define VERSION "6-linux-konica"
+
+#endif  // BRLASER_CONFIG_H